set up user authorization and policies for CRUD with the listings

+++ app/Http/Controllers/Admin/ListingController.php
@@ -17,7 +17,7 @@ class ListingController extends Controller
      */
     public function index()
     {
-        $listings = Listing::paginate(2);
+        $listings = Listing::where('user_id', auth()->user()->id)->paginate(2);
 
         return view('admin/listings/index', [
             'listings' => $listings
@@ -31,6 +31,7 @@ class ListingController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Listing::class);
         return view('admin/listings/create');
     }
 
@@ -42,6 +43,8 @@ class ListingController extends Controller
      */
     public function store(Request $request)
     {
+      $this->authorize('create', Listing::class);
+
       request()->validate([
         'address' => 'required',
         'city' => 'required',
@@ -52,6 +55,7 @@ class ListingController extends Controller
       ]);
         
         $listing = new Listing();
+        $listing->user_id = auth()->user()->id;
         $listing->address = $request->get('address');
         $listing->address2 = $request->get('address2');
         $listing->city = $request->get('city');
@@ -93,6 +97,7 @@ class ListingController extends Controller
             'slug' => $slug
         ])->first();
         
+        $this->authorize('update', $listing);
         return view('admin/listings/edit', ['listing' => $listing]);
     }
 
@@ -118,6 +123,8 @@ class ListingController extends Controller
             'id' => $id,
             'slug' => $slug
         ])->first();
+
+        $this->authorize('update', $listing);
 
         $listing->address = $request->get('address');
         $listing->address2 = $request->get('address2');
@@ -146,6 +153,8 @@ class ListingController extends Controller
     {
         $listing = Listing::find($id);
 
+        $this->authorize('delete', $listing);
+        
         $listing->delete();
 
         return redirect("/admin/listings")->with('success', 'Listing Has Been Deleted Successfully');
